v3.2

+++ src/main/java/org/example/controllers/TasksController.java
@@ -1,6 +1,7 @@
 package org.example.controllers;
 
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Min;
 import org.example.DTO.*;
 import org.example.enums.Status;
 import org.example.models.Task;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.*;
 
 import lombok.RequiredArgsConstructor;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -39,14 +41,14 @@ public class TasksController {
             Optional<User> userOptional = userService.getUserById(addTaskRequest.getUser_id());
             if (userOptional.isPresent()) {
                 User user = userOptional.get();
-                task.setUser(user); // Связываем задачу с пользователем
+                task.setUser(user);
             } else {
-                return new ResponseEntity<>("Пользователь с ID " + addTaskRequest.getUser_id() + " не найден", HttpStatus.NOT_FOUND);
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
             }
             Task addedTask = taskService.addTask(task);
-            return new ResponseEntity<>(addedTask.getId(), HttpStatus.CREATED); // Вернуть ID и код 201
+            return new ResponseEntity<>(addedTask.getId(), HttpStatus.CREATED); //201
         } catch (Exception e) {
-            return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR); //500
         }
     }
 
@@ -63,66 +65,93 @@ public class TasksController {
             if (changedTask.isPresent()) {
                 return new ResponseEntity<>(changedTask.get(), HttpStatus.OK);
             } else {
-                return new ResponseEntity<>("Задача не найдена или не принадлежит пользователю", HttpStatus.NOT_FOUND);
+                return new ResponseEntity<>("Задача не найдена или не принадлежит пользователю", HttpStatus.NOT_FOUND); //404
             }
         } catch (Exception e) {
-            return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR); //500
         }
     }
 
     @PutMapping("/completeTask/{taskId}/{userId}")
-    public ResponseEntity<?> completeTask(@PathVariable Long taskId, @PathVariable Long userId) {
+    public ResponseEntity<?> completeTask(@PathVariable @Min(value = 0, message = "ID задачи не может быть <0") Long taskId, @PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             taskService.completeTask(taskId, userId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // Код 204
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); //204
         } catch (Exception e) {
-            return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR); //500
         }
     }
 
     @PutMapping("/returnCompletedTask/{taskId}/{userId}")
-    public ResponseEntity<?> returnCompletedTask(@PathVariable Long taskId, @PathVariable Long userId) {
+    public ResponseEntity<?> returnCompletedTask(@PathVariable @Min(value = 0, message = "ID задачи не может быть <0") Long taskId, @PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             taskService.returnCompletedTask(taskId, userId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // Код 204
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); //204
         } catch (Exception e) {
             return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/deleteCompletedTasks/{userId}")
-    public ResponseEntity<?> deleteCompletedTasks(@PathVariable Long userId) {
+    public ResponseEntity<?> deleteCompletedTasks(@PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             taskService.deleteCompletedTasks(userId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // Код 204
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); //204
         } catch (Exception e) {
             return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/deleteTask/{taskId}/{userId}")
-    public ResponseEntity<?> deleteTask(@PathVariable Long taskId, @PathVariable Long userId) {
+    public ResponseEntity<?> deleteTask(@PathVariable @Min(value = 0, message = "ID задачи не может быть <0") Long taskId, @PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             taskService.deleteTask(taskId, userId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // Код 204
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); //204
         } catch (Exception e) {
             return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping("/getAllTasks/{userId}")
-    public ResponseEntity<?> getAllTasks(@PathVariable Long userId) {
+    public ResponseEntity<?> getAllTasks(@PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             List<Task> tasks = taskService.getAllTasks(userId);
-            return new ResponseEntity<>(tasks, HttpStatus.OK);
+            List<GetAllTasksResponse> listResult = new ArrayList<>();
+            for (Task task : tasks){
+                GetAllTasksResponse getAllTasksResponse = new GetAllTasksResponse();
+                getAllTasksResponse.setId(task.getId());
+                getAllTasksResponse.setTitle(task.getTitle());
+                getAllTasksResponse.setDate(task.getDate());
+                getAllTasksResponse.setTime(task.getTime());
+                listResult.add(getAllTasksResponse);
+            }
+            return new ResponseEntity<>(listResult, HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>("Ошибка: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping("/getCompletedTasks/{userId}")
-    public ResponseEntity<?> getCompletedTasks(@PathVariable Long userId) {
+    public ResponseEntity<?> getCompletedTasks(@PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             List<Task> tasks = taskService.getCompletedTasks(userId);
             return new ResponseEntity<>(tasks, HttpStatus.OK);
         } catch (Exception e) {
@@ -131,8 +160,11 @@ public class TasksController {
     }
 
     @GetMapping("/getTodayTasks/{userId}")
-    public ResponseEntity<?> getTodayTasks(@PathVariable Long userId) {
+    public ResponseEntity<?> getTodayTasks(@PathVariable @Min(value = 0, message = "ID пользователя не может быть <0") Long userId) {
         try {
+            if (userService.getUserById(userId).isEmpty()) {
+                return new ResponseEntity<>("Пользователь не найден", HttpStatus.NOT_FOUND); //404
+            }
             List<Task> tasks = taskService.getTodayTasks(userId);
             return new ResponseEntity<>(tasks, HttpStatus.OK);
         } catch (Exception e) {
